+++ src/midje_nrepl/misc.clj
@@ -1,5 +1,6 @@
 (ns midje-nrepl.misc
-  (:require             [orchard.classpath :as classpath]))
+  (:require [orchard.classpath :as classpath])
+  (:import (java.time Duration Instant)))
 
 (defn dependency-in-classpath?
   "Return true if a given dependency is in the project's classpath, or false otherwise."
@@ -19,5 +20,14 @@
   (reduce (fn [result [key parse-fn]]
             (if-let [value (get options key)]
               (assoc result key (parse-fn value))
-              result)
-            ) {} parsers-map))
+              result)) {} parsers-map))
+
+(defn now
+  "Return a java.time.Instant object representing the current instant."
+  []
+  (Instant/now))
+
+(defn duration-between
+  "Return a java.time.Duration object representing the duration between two temporal objects."
+  [start end]
+  (Duration/between start end))