+++ test/midje_nrepl/middleware/test_test.clj
@@ -3,9 +3,11 @@
             [clojure.tools.nrepl.transport :as transport]
             [matcher-combinators.midje :refer [match]]
             [midje-nrepl.middleware.test :as test]
-            [midje-nrepl.test-runner :as test-runner]
+            [midje-nrepl.misc :as misc]
+            [midje-nrepl.runner :as runner]
             [midje.sweet :refer :all]
-            [orchard.misc :as misc]))
+            [orchard.misc :refer [transform-value]])
+  (:import java.time.Duration))
 
 (def test-report {:results
                   {'octocat.arithmetic-test
@@ -16,20 +18,24 @@
                      :expected 6
                      :actual   5
                      :message  '()
-                     :type     :fail}]}
+                     :type     :fail
+                     :started-at (misc/now)
+                     :finished-at (misc/now)}]}
                   :summary {:check 1 :error 0 :fact 1 :fail 1 :ns 1 :pass 0 :to-do 0}})
 
-(def transformed-report (misc/transform-value test-report))
+(def transformed-report (assoc-in (transform-value test-report) ["summary" "finished-in"] "1 millisecond"))
 
 (def exception (RuntimeException. "An unexpected error was thrown" (ArithmeticException. "Divid by zero")))
 
 (facts "about handling test operations"
+       (against-background
+        (misc/duration-between anything anything) => (.plusMillis (Duration/ZERO) 1))
 
        (fact "run all tests in the project and sends the report to the client"
              (test/handle-test {:op        "midje-test-all"
                                 :transport ..transport..}) => irrelevant
              (provided
-              (test-runner/run-all-tests {}) => test-report
+              (runner/run-all-tests {}) => test-report
               (transport/send ..transport.. transformed-report) => irrelevant
               (transport/send ..transport.. (match {:status #{:done}})) => irrelevant))
 
@@ -38,7 +44,7 @@
                                 :test-paths ["src/clojure/test"]
                                 :transport  ..transport..}) => irrelevant
              (provided
-              (test-runner/run-all-tests {:test-paths ["src/clojure/test"]}) => test-report
+              (runner/run-all-tests {:test-paths ["src/clojure/test"]}) => test-report
               (transport/send ..transport.. transformed-report) => irrelevant
               (transport/send ..transport.. (match {:status #{:done}})) => irrelevant))
 
@@ -48,9 +54,19 @@
                                 :ns-inclusions ["^integration"]
                                 :transport     ..transport..}) => irrelevant
              (provided
-              (test-runner/run-all-tests (match {:ns-exclusions #(= (map str %) ["^integration\\.too-heavy"])
-                                                 :ns-inclusions #(= (map str %) ["^integration"])})) => test-report
+              (runner/run-all-tests (match {:ns-exclusions #(= (map str %) ["^integration\\.too-heavy"])
+                                            :ns-inclusions #(= (map str %) ["^integration"])})) => test-report
               (transport/send ..transport.. transformed-report) => irrelevant
+              (transport/send ..transport.. (match {:status #{:done}})) => irrelevant))
+
+       (fact "clients can collect profiling information by sending the parameter `profile?`"
+             (test/handle-test {:op        "midje-test-all"
+                                :profile?  "true"
+                                :transport ..transport..}) => irrelevant
+             (provided
+              (runner/run-all-tests {:profile? true}) => test-report
+              (transport/send ..transport.. (contains
+                                             (assoc transformed-report "profile" anything))) => irrelevant
               (transport/send ..transport.. (match {:status #{:done}})) => irrelevant))
 
        (fact "runs all tests in the given namespace and sends the report to the client"
@@ -58,7 +74,7 @@
                                 :ns        "octocat.arithmetic-test"
                                 :transport ..transport..}) => irrelevant
              (provided
-              (test-runner/run-tests-in-ns 'octocat.arithmetic-test) => test-report
+              (runner/run-tests-in-ns {:ns 'octocat.arithmetic-test}) => test-report
               (transport/send ..transport.. transformed-report) => irrelevant
               (transport/send ..transport.. (match {:status #{:done}})) => irrelevant))
 
@@ -69,7 +85,7 @@
                                 :source    "(fact (+ 2 3) => 6)"
                                 :transport ..transport..}) => irrelevant
              (provided
-              (test-runner/run-test 'octocat.arithmetic-test "(fact (+ 2 3) => 6)" 10) => test-report
+              (runner/run-test {:ns 'octocat.arithmetic-test :source "(fact (+ 2 3) => 6)" :line 10}) => test-report
               (transport/send ..transport.. transformed-report) => irrelevant
               (transport/send ..transport.. (match {:status #{:done}})) => irrelevant))
 
@@ -77,7 +93,7 @@
              (test/handle-test {:op        "midje-retest"
                                 :transport ..transport..}) => irrelevant
              (provided
-              (test-runner/re-run-non-passing-tests) => test-report
+              (runner/re-run-non-passing-tests {}) => test-report
               (transport/send ..transport.. transformed-report) => irrelevant
               (transport/send ..transport.. (match {:status #{:done}})) => irrelevant))
 
@@ -88,7 +104,7 @@
                                 :print-fn  println
                                 :transport ..transport..}) => irrelevant
              (provided
-              (test-runner/get-exception-at 'octocat.arithmetic-test 2) => exception
+              (runner/get-exception-at 'octocat.arithmetic-test 2) => exception
               (transport/send ..transport.. (match {:class      "java.lang.RuntimeException"
                                                     :message    "An unexpected error was thrown"
                                                     :stacktrace (complement empty?)}))  => irrelevant
@@ -104,6 +120,6 @@
                                 :print-fn  println
                                 :transport ..transport..}) => irrelevant
              (provided
-              (test-runner/get-exception-at 'octocat.arithmetic-test 0) => nil
+              (runner/get-exception-at 'octocat.arithmetic-test 0) => nil
               (transport/send ..transport.. (match {:status #{:no-stacktrace}})) => irrelevant
               (transport/send ..transport.. (match {:status #{:done}})) => irrelevant)))