+++ test/midje_nrepl/runner_test.clj
@@ -1,9 +1,9 @@
-(ns midje-nrepl.test-runner-test
-  (:require [matcher-combinators.midje :refer [match]]
+(ns midje-nrepl.runner-test
+  (:require [matcher-combinators.matchers :as m]
+            [matcher-combinators.midje :refer [match]]
             [midje-nrepl.project-info :as project-info]
-            [midje-nrepl.test-runner :as test-runner]
-            [midje.sweet :refer :all]
-            [matcher-combinators.matchers :as m]))
+            [midje-nrepl.runner :as runner]
+            [midje.sweet :refer :all]))
 
 (defn existing-file? [candidate]
   (and   (instance? java.io.File candidate)
@@ -154,7 +154,7 @@
 (defn isolate-test-forms!
   "Workaround to test the re-run feature without modifying Midje counters."
   [namespace]
-  (swap! test-runner/test-results update namespace
+  (swap! runner/test-results update namespace
          (fn [results]
            (->> results
                 (map #(update % :source
@@ -164,62 +164,62 @@
 (facts "about running tests in a given namespace"
 
        (fact "runs all tests in the given namespace"
-             (test-runner/run-tests-in-ns 'octocat.arithmetic-test) => (match arithmetic-test-report)
-             (test-runner/run-tests-in-ns 'octocat.colls-test) => (match colls-test-report)
-             (test-runner/run-tests-in-ns 'octocat.mocks-test) => (match mocks-test-report))
+             (runner/run-tests-in-ns {:ns 'octocat.arithmetic-test}) => (match arithmetic-test-report)
+             (runner/run-tests-in-ns {:ns 'octocat.colls-test}) => (match colls-test-report)
+             (runner/run-tests-in-ns {:ns 'octocat.mocks-test}) => (match mocks-test-report))
 
        (fact "returns a report with no tests when there are no tests to be run"
-             (test-runner/run-tests-in-ns 'octocat.no-tests)
+             (runner/run-tests-in-ns {:ns 'octocat.no-tests})
              => (match {:results {}
                         :summary {:check 0 :ns 0}}))
 
        (fact "results of the last execution are kept in the current session"
-             (test-runner/run-tests-in-ns 'octocat.arithmetic-test) => (match arithmetic-test-report)
-             @test-runner/test-results
+             (runner/run-tests-in-ns {:ns 'octocat.arithmetic-test}) => (match arithmetic-test-report)
+             @runner/test-results
              => (match (:results arithmetic-test-report))))
 
 (facts "about running individual tests"
 
        (fact "runs the test source passed as a string"
-             (test-runner/run-test 'octocat.arithmetic-test "(with-isolated-output-counters (fact 1 => 1))")
+             (runner/run-test {:ns 'octocat.arithmetic-test :source "(with-isolated-output-counters (fact 1 => 1))"})
              => (match individual-test-report))
 
        (fact "line numbers are resolved correctly for individual facts, taking the supplied starting line in consideration"
-             (test-runner/run-test 'octocat.arithmetic-test "(with-isolated-output-counters
+             (runner/run-test {:ns 'octocat.arithmetic-test :source "(with-isolated-output-counters
 (fact \"this is wrong\"
-1 => 2))" 10)
+1 => 2))"                          :line                    10})
              => (match {:results {'octocat.arithmetic-test
                                   [{:type :fail
                                     :line 12}]}}))
 
        (fact "returns a report with no tests when there are no tests to be run"
-             (test-runner/run-test 'octocat.arithmetic-test "(fact)")
+             (runner/run-test {:ns 'octocat.arithmetic-test :source "(fact)"})
              => (match {:results {}
                         :summary {:check 0 :ns 0}}))
 
        (fact "results of the last execution are kept in the current session"
-             (test-runner/run-test 'octocat.arithmetic-test "(with-isolated-output-counters (fact 1 => 1))")
+             (runner/run-test {:ns 'octocat.arithmetic-test :source "(with-isolated-output-counters (fact 1 => 1))"})
              => (match individual-test-report)
-             @test-runner/test-results
+             @runner/test-results
              => (match (:results individual-test-report))))
 
 (facts "about running all tests in the project"
 
        (fact "runs all tests in the project according to supplied options"
-             (test-runner/run-all-tests {:test-paths ["test/octocat"]})
+             (runner/run-all-tests {:test-paths ["test/octocat"]})
              => (match all-tests-report))
 
        (fact "when `test-paths` isn't set, uses the test paths declared in the project"
-             (test-runner/run-all-tests {})
+             (runner/run-all-tests {})
              => (match all-tests-report)
              (provided
               (project-info/get-test-paths) => ["test/octocat"]))
 
        (tabular (fact "`:ns-exclusions` and `:ns-inclusions` allow for users to
               filter out the list of namespaces to be tested"
-                      (-> (test-runner/run-all-tests {:test-paths    ["test/octocat"]
-                                                      :ns-exclusions ?exclusions
-                                                      :ns-inclusions ?inclusions})
+                      (-> (runner/run-all-tests {:test-paths    ["test/octocat"]
+                                                 :ns-exclusions ?exclusions
+                                                 :ns-inclusions ?inclusions})
                           :results
                           keys)
                       => ?results)
@@ -233,60 +233,60 @@
 
        (fact "when both `:ns-exclusions` and `:ns-inclusions` are present, the
        former takes precedence over the later"
-             (test-runner/run-all-tests {:test-paths    ["test/octocat"]
-                                         :ns-exclusions [#"arithmetic"]
-                                         :ns-inclusions [#"arithmetic"]})
+             (runner/run-all-tests {:test-paths    ["test/octocat"]
+                                    :ns-exclusions [#"arithmetic"]
+                                    :ns-inclusions [#"arithmetic"]})
              => (match {:results empty?}))
 
        (fact "returns a report with no tests when there are no tests to be run"
-             (test-runner/run-all-tests {:test-paths ["test/octocat"]})
+             (runner/run-all-tests {:test-paths ["test/octocat"]})
              => (match {:results {}
                         :summary {:check 0 :ns 0}})
              (provided
               (project-info/find-namespaces-in ["test/octocat"]) => ['octocat.no-tests]))
 
        (fact "results of the last execution are kept in the current session"
-             (test-runner/run-all-tests {:test-paths ["test/octocat"]})
+             (runner/run-all-tests {:test-paths ["test/octocat"]})
              => (match all-tests-report)
-             @test-runner/test-results
+             @runner/test-results
              => (match (:results all-tests-report))))
 
 (facts "about re-running tests"
 
        (fact "re-runs tests that didn't pass in the last execution"
-             (test-runner/run-tests-in-ns 'octocat.arithmetic-test) => (match arithmetic-test-report)
+             (runner/run-tests-in-ns {:ns 'octocat.arithmetic-test}) => (match arithmetic-test-report)
              (isolate-test-forms! 'octocat.arithmetic-test)
-             (test-runner/re-run-non-passing-tests) => (match re-run-arithmetic-test-report))
+             (runner/re-run-non-passing-tests nil) => (match re-run-arithmetic-test-report))
 
        (fact "returns a report with no tests when there are no failing or erring tests to be run"
-             (test-runner/run-test 'octocat.arithmetic-test "(with-isolated-output-counters (fact (+ 1 2) => 3))")
+             (runner/run-test {:ns 'octocat.arithmetic-test :source "(with-isolated-output-counters (fact (+ 1 2) => 3))"})
              => (match {:summary {:error 0 :fail 0 :ns 1 :pass 1 :check 1}})
-             (test-runner/re-run-non-passing-tests)
+             (runner/re-run-non-passing-tests nil)
              => (match {:results {}
                         :summary {:check 0 :ns 0}}))
 
        (fact "results of the last execution are kept in the current session as well"
-             (test-runner/run-tests-in-ns 'octocat.arithmetic-test) => (match arithmetic-test-report)
+             (runner/run-tests-in-ns {:ns 'octocat.arithmetic-test}) => (match arithmetic-test-report)
              (isolate-test-forms! 'octocat.arithmetic-test)
-             (test-runner/re-run-non-passing-tests) => (match re-run-arithmetic-test-report)
-             @test-runner/test-results
+             (runner/re-run-non-passing-tests nil) => (match re-run-arithmetic-test-report)
+             @runner/test-results
              => (match (:results re-run-arithmetic-test-report)))
 
        (fact "by re-running the same tests twice, the same results are obtained"
-             (test-runner/run-tests-in-ns 'octocat.arithmetic-test) => (match arithmetic-test-report)
+             (runner/run-tests-in-ns {:ns 'octocat.arithmetic-test}) => (match arithmetic-test-report)
              (isolate-test-forms! 'octocat.arithmetic-test)
-             (test-runner/re-run-non-passing-tests) => (match re-run-arithmetic-test-report)
+             (runner/re-run-non-passing-tests nil) => (match re-run-arithmetic-test-report)
              (isolate-test-forms! 'octocat.arithmetic-test)
-             (test-runner/re-run-non-passing-tests) => (match re-run-arithmetic-test-report)))
+             (runner/re-run-non-passing-tests nil) => (match re-run-arithmetic-test-report)))
 
 (facts "about getting test stacktraces"
 
        (fact "given a namespace symbol and an index, returns the exception stored at these coordinates"
-             (test-runner/run-tests-in-ns 'octocat.arithmetic-test) => irrelevant
-             (test-runner/get-exception-at 'octocat.arithmetic-test 4) => #(instance? ArithmeticException %))
+             (runner/run-tests-in-ns {:ns 'octocat.arithmetic-test}) => irrelevant
+             (runner/get-exception-at 'octocat.arithmetic-test 4) => #(instance? ArithmeticException %))
 
        (tabular (fact "when there is no exception at the supplied position, returns nil"
-                      (test-runner/get-exception-at ?ns ?index) => nil)
+                      (runner/get-exception-at ?ns ?index) => nil)
                 ?ns                       ?index
                 'octocat.arithmetic-test       0
                 'octocat.arithmetic-test      10